Share Screen: Saving output of the translator into the database, Preparation of the SubmissionData and UploadData for further upload. Error handling. New DB Request - CreateItemWithAttachmentDbRequest

+++ app/src/main/java/org/zotero/android/backgrounduploader/BackgroundUpload.kt
@@ -8,7 +8,7 @@ data class BackgroundUpload(
     val type: Kind,
     val key: String,
     val libraryId: LibraryIdentifier,
-    val userId: Int,
+    val userId: Long,
     val remoteUrl: URL,
     val fileUrl: URL,
     val md5: String,
@@ -18,7 +18,7 @@ data class BackgroundUpload(
 ) {
     sealed class Kind {
         data class zotero(val uploadKey: String): Kind()
-        data class webdav(val mtime: Int): Kind()
+        data class webdav(val mtime: Long): Kind()
     }
 
 }
